pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ducky-network-tool"
version = "1.3.0"
authors = [
  { name="Your Name", email="youremail@example.com" },
]
description = "An all-in-one toolkit for network engineers and security enthusiasts."
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "PySide6",
    "pyserial",
    "psutil",
    "scapy",
    "requests",
    "zxcvbn",
    "paramiko",
    "telnetlib3",
    "pysnmp"
]

[project.scripts]
ducky = "ducky_app.main:run"

[project.urls]
Homepage = "https://github.com/thecmdguy/Ducky"

[tool.setuptools]
package-dir = {"" = "src"}
packages = {find = {where=["src"]}}